// src/service/file/schemas/studentRecordResult.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { StudentRecordResultCourse } from './studentRecordResultCourse';
import type { StudentRecordEvaluationIndicatorScoreResult } from './studentRecordEvaluationIndicatorScoreResult';
import type { StudentRecordAttachedPhotoResult } from './studentRecordAttachedPhotoResult';
import type { StudentRecordCommentResult } from './studentRecordCommentResult';
import type { DriveItemResult } from './driveItemResult';
import type { SmartFolderItemResult } from './smartFolderItemResult';

export interface StudentRecordResult {
  /** 관찰기록의 아이디 */
  id: number;
  /**
   * 작성하는 선생님의 계정 아이디
   * @minimum 0
   */
  creatorAccountId: number;
  /**
   * 작성하는 선생님의 프로필 아이디
   * @minimum 0
   */
  creatorProfileId: number;
  /**
   * 아이가 속한 반의 아이디
   * @minimum 0
   */
  educationalClassId: number;
  /** 아이가 속한 반의 연도 */
  educationalClassYear?: number;
  /** 아이가 속한 반의 이름 */
  educationalClassName?: string;
  /** 아이가 속한 반의 연령 */
  educationalClassAge?: number;
  /** 아이의 아이디 */
  studentId: number;
  /** 아이의 이름 */
  studentName?: string;
  /** 아이의 사진 썸네일 */
  studentThumbnail?: string;
  /** 아이의 생년월일 */
  studentBirthday?: string;
  /** 아이가 현재 해당하는 교육 과정 */
  course: StudentRecordResultCourse;
  /**
   * 방사형 그래프에 영역과 평균 점수
   * @maxLength 30
   */
  summaryScores?: string;
  /**
   * 대표사진 url
   * @maxLength 1000
   */
  thumbUrl?: string;
  /**
   * 한줄평 이미지
   * @maxLength 1000
   */
  typeImageUrl?: string;
  /**
   * 한줄평 내용
   * @maxLength 200
   */
  typeDescription?: string;
  /** 생성일 */
  createdAt: string;
  /** 수정일 */
  modifiedAt: string;
  /** 그래프에 들어간 항목에 대한 점수들 */
  indicatorScores?: StudentRecordEvaluationIndicatorScoreResult[];
  /** 첨부된 사진들 */
  attachedPhotos?: StudentRecordAttachedPhotoResult[];
  /** 관찰 평가 영역별 내용들 */
  observeComments?: StudentRecordCommentResult[];
  /** 교사지원 내용 */
  teacherSupport?: string;
  /** 교사지원 내용을 사용자가 직접 수정 했는지? */
  teacherSupportUserEdited?: boolean;
  /** 학부모 지원 내용 */
  parentSupport?: string;
  /** 학부모 지원 내용을 사용자가 직접 수정 했는지? */
  parentSupportUserEdited?: boolean;
  /** 교사 총평 내용 */
  teacherComment?: string;
  /** 교사 총평 내용을 사용자가 직접 수정 했는지? */
  teacherCommentUserEdited?: boolean;
  /** 관찰 요약 내용 */
  observeSummary?: string;
  /** 관찰 요약 내용을 사용자가 직접 수정 했는지? */
  observeSummaryUserEdited?: boolean;
  driveItem?: DriveItemResult;
  smartFolderItem?: SmartFolderItemResult;
}